Refactor API util
